Remove cmd-list.c line in master.
